requirements.txt
# Generated by vnm
profanity-filter==1.3.3
pybuildtools==0.4.2
werkzeug==1.0.1
ruamel.yaml==0.15.100
